+++ updatable_message.go
